Move fork directory to stdlib
