+++ runtime-light/component/component.h
@@ -14,9 +14,9 @@
 #include "runtime-core/runtime-core.h"
 #include "runtime-light/core/globals/php-script-globals.h"
 #include "runtime-light/coroutine/task.h"
-#include "runtime-light/fork/fork-context.h"
 #include "runtime-light/header.h"
 #include "runtime-light/scheduler/scheduler.h"
+#include "runtime-light/stdlib/fork/fork-context.h"
 #include "runtime-light/stdlib/output-control.h"
 #include "runtime-light/stdlib/rpc/rpc-context.h"
