+++ runtime-light/coroutine/awaitable.h
@@ -15,8 +15,8 @@
 #include "runtime-core/utils/kphp-assert-core.h"
 #include "runtime-light/component/component.h"
 #include "runtime-light/coroutine/task.h"
-#include "runtime-light/fork/fork-context.h"
-#include "runtime-light/fork/fork.h"
+#include "runtime-light/stdlib/fork/fork-context.h"
+#include "runtime-light/stdlib/fork/fork.h"
 #include "runtime-light/header.h"
 #include "runtime-light/scheduler/scheduler.h"
 #include "runtime-light/utils/context.h"
