--- runtime-light/fork/fork.cmake
@@ -1,1 +1,0 @@
-prepend(RUNTIME_FORK_SRC fork/ fork-api.cpp fork-context.cpp)
