+++ runtime-light/stdlib/fork/fork-api.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2024 LLC «V Kontakte»
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-#include "runtime-light/fork/fork-api.h"
+#include "runtime-light/stdlib/fork/fork-api.h"
 
 #include <chrono>
 #include <cstdint>
