+++ runtime-light/stdlib/stdlib.cmake
@@ -1,15 +1,18 @@
-prepend(RUNTIME_STDLIB_SRC stdlib/
-        interface.cpp
-        misc.cpp
-        output-control.cpp
-        string-functions.cpp
-        variable-handling.cpp
-        superglobals.cpp
-        rpc/rpc-api.cpp
-        rpc/rpc-context.cpp
-        rpc/rpc-extra-headers.cpp
-        rpc/rpc-extra-info.cpp
-        rpc/rpc-tl-error.cpp
-        rpc/rpc-tl-query.cpp
-        rpc/rpc-tl-request.cpp
-)
+prepend(
+  RUNTIME_STDLIB_SRC
+  stdlib/
+  interface.cpp
+  misc.cpp
+  output-control.cpp
+  string-functions.cpp
+  variable-handling.cpp
+  superglobals.cpp
+  fork/fork-api.cpp
+  fork/fork-context.cpp
+  rpc/rpc-api.cpp
+  rpc/rpc-context.cpp
+  rpc/rpc-extra-headers.cpp
+  rpc/rpc-extra-info.cpp
+  rpc/rpc-tl-error.cpp
+  rpc/rpc-tl-query.cpp
+  rpc/rpc-tl-request.cpp)
